+++ src/components/VisualEditor/VisualEditor.test.tsx
@@ -1,7 +1,7 @@
 import { toDataFrame } from '@grafana/data';
 import { DragDropContext, Draggable, Droppable } from '@hello-pangea/dnd';
 import { fireEvent, render, screen } from '@testing-library/react';
-import { getJestSelectors } from '@volkovlabs/jest-selectors';
+import { getJestSelectors } from '../../test-utils/jest-selectors';
 import React from 'react';
 
 import { TEST_IDS } from '../../constants';